Make `configure_misses` output more reproducible
Because there might be multiple `configure` scripts running in parallel during
the build, the scripts' output could be interleaved, leaving the
`configure_misses` output in an unpredictable state.

Use `make -O ...` to eliminate the interleaving: The output will contain
predictable "chunks". Output from each `configure` script will be contained in
one of the chunks, separate from one another. The ordering of the chunks is not
predictable, but because `configure_misses` is sorted in the end, the committed
results should be reproducible.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>